service_resultをmodule化
